Prevent audio output test from hanging once data has run out.
This test feeds exactly 2 seconds of audio data to an AudioOutputStream
over a SyncSocket. Unfortunately, the audio producer is shut down
unceremoniously without signaling to the consumer that no more audio
data will be sent. This leads to a race condition, where the test is
trying to stop the AudioOutputStream before it attempts to read more
data from the SyncSocket. When we lose the race, the test hangs, waiting
for data that will never come.

The solution is to enlighten the consumer in this test to expect 2
seconds of audio data, and have it stop requesting more once it has
reached the expected number of audio frames.

Bug: 144021
Change-Id: I158023d1831698e2e642a8d3fab518ce35b98154
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157777
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Olga Sharonova <olka@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#761856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 604dfe11c1488ab0e9e2179d173d7b764f15dded