Fix race condition between Resume / TrackChange
As can be seen in the test case, the pipeline controller can get a
resume event after the track change has been sent to the pipeline,
but before the pipeline has responded. In that case, the resume
event would just be dropped. This change causes the resume to stick
around in that one specific case so that the track change can
re-evaluate it's necessity afterwards.

Thanks to marcel333mail@gmail.com for the testcase.

Bug: 1072230
Change-Id: Ia88beb51cde085632346b1b5ba930af4d971337f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163866
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#762614}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 618e7929acf23f0126f73ac9bf17658546b99e0a