vaapi: Fix profile 1 VP9 hw encode related issues
We see that even though VP9 HW encode on profile 1 is not enabled
on chromium-vaapi, it is providing options to use it based on
media driver's capabilities. We observed an issue with this
while using apprtc application to do VP9 encode/decode loopback call.
The application is requesting for supported profiles during VP9 HW
encode and chromium-vaapi is providing the profile1 option that media
driver supports and it not enabled on chromium-vaapi. During the
initiation of the call, VP9 hw encode fallsback to software encode.
This workaround will hide profile 1 during VP9 hw encode which makes
the applications to use profile 0 and continue to encode using hardware.

BUG = b:153680337
TEST = https://appr.tc/?debug=loopback&vrc=vp9&vsc=vp9

Signed-off-by: Sushma Venkatesh Reddy <sushma.venkatesh.reddy@intel.com>
Change-Id: I111f91a0ce58e36d35d995bbc7328805c163deef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2145079
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Sreerenj Balachandran <sreerenj.balachandran@intel.com>
Cr-Original-Commit-Position: refs/heads/master@{#762955}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9534c98ca89cf67592883ba0e1bb4d9f5e78145d