Avoid calling into MediaCodec during startup; assume support.
The prevents us from calling into MediaCodec during GPU process
startup by assuming support for H264 hardware decode and encode as
well as VP8 encode (since we now only support Lollipop+).

The outcome of this change is that we may later find that H264 or
VP8 are not hardware accelerated. If an encoder is created before
the real list comes back from the VEA, it will fail to initialize.

In the case of VP8, this will just cause fallback to an internal
software encoder. In the rare case of H264 not being accelerated
this will likely trigger an error. VP8 is used in 92% of cases,
so this should be rare, but we can watch these metrics for issues:
https://uma.googleplex.com/p/chrome/histograms/?sid=b3aa4846b1d0cc3f715c86eae320e08a

RenderThreadImpl::GetGpuFactories() is always called early in the
lifetime of a WebRTC stream and used asynchronously AFAICT, so in
the vast majority of cases we should have the real information
ahead of it actually being used.

Fixed: 709631, 1060929
Change-Id: Ic46f9a3e37cbd857945f5cc03b7619fae923c1f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165723
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#763079}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3f8ba374d20edb3890e716abf24e5dd6417fd8a