Use SilentSinkSuspender for normal audio playback.
This adds a SilentSinkSuspender to AudioRendererMixer so that silent
audio during a normal playback ends up disabling the physical audio
stream for a significant amount of power saving when possible.

The plan is to land this behind a feature flag and then run some
power tests on Canary to see if this is worth enabling by default.

This doesn't disable the audio track, so we do still decode the
audio, but generally we have to in case the playback is attached
for WebAudio. As a first step lets see how effective this is.

R=olka, tmathmeyer

Bug: 764133
Change-Id: I3d063eaf80e03bed38f3ba002c5ab5813b79fe83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194915
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Ted Meyer <tmathmeyer@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#767976}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 314b89bc686b337430a4f85919417dc0064a87b4