Add special case for |FFmpegDemuxer::SeekInternal| with -INF start_time_
base::TimeDelta will fail a DCHECK if you ever try to add -INF and +INF
time deltas. A consequence of this is if you have a video with a -INF
start_time and you try to seek to +INF, the current code will trigger
that DCHECK. This CL fixes that by settings -INF start times to 0.
This is consistent with how +INF start times are handled.

Bug: 1077685
Change-Id: I4e7fd453aab7a6b15befbb52b708d56febb943ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2194917
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Will Cassella <cassew@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#770856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 67699d870caf26f19ef17dccad13840d52e27de3