media/gpu/test: Enable/Disable features in a test
There are some media features that needs to be overridden in
a test. For instance, we want to enable kVaapiVP9Encoder, which
is disabled by default now, and also kFFmpegDecodeOpaque to use
VpxVideoDecoder to decode vp8 bitstream for a verification.

There is no way to enable and disable features in our video
tests. After this CL, media::test::VideoTestEnvironment takes
feature sets to be either enabled or disabled in a test.

Bug: 1045825
Test: video_encode_accelerator_tests --codec=vp8 on eve
Change-Id: Ib66d96c124783ca47c3e943d396e6666ebac03ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2206872
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#771347}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a69c1dc6a274f4bb6492a321df6a6e421ebfb844