media/filter: Remove GetMaxDecodeRequests from gav1_video_decoder
GetMaxAllowedFrames() in libgav1 is going to be removed from the
public API. It does not return any meaningful values today. Some
of the other wrappers don't use it as well, so it should be ok to
remove it.

Bug: 1047051
Change-Id: I4b7eb14d25e8d809a54c2114f3eb758ac163a5ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220354
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: James Zern <jzern@google.com>
Commit-Queue: James Zern <jzern@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#772870}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 285e6deb9e4c9e68fd5ce6647004c8a5d4aaf8d2