media/gpu/v4l2vd: support finer-grained control of queues streaming
The stateless V4L2 decoder, which was the only V4L2 decoder
supported so far, allowed both queues to be started and stopped at
the same time. This is not the case for the stateful decoder, which
need to start the input queue first, and to keep it running when
changing resolution. Add a parameter to the streaming control
methods of the V4L2VD to allow this behavior to take place, as well
as an extra method in the backend interface to let the VD know
whether it should stop the input queue or not during resolution
changes.

This decoupling of the queues means that the input queue may now
have buffers queued when we change resolution. So remove the
DCHECK() that checked against this since it is obsolete.

BUG=b:149663704
TEST=VDAtest --use_vd passes on Krane.

Change-Id: I038b61ced2f09f7d390be1716117c65c245a10bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235581
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#776507}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9270ddceb3282beab94abf6767cb69cf05f88678