media: Prevent int overflow during GMB allocation.
This CL modifies the helper function AllocateGpuMemoryBufferHandle() to
prevent an int overflows when getting the ID of the next GpuMemoryBuffer
to be allocated.

If we are about to overflow, we just crash the GPU process. This is not
the ideal situation, but I don't expect it to happen often. There are
other solutions but they would be more intrusive.

Bug: 1092283
Test: video.Play.h264 on eve
Change-Id: Iba28a996837b5d37ab42c14e457c6e1d54b6b3cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238870
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#777031}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef9c8ce2c40fe72b30db5836045471700fa6ccc8