media/gpu/v4l2vd: start streaming input queue during Initialize()
The stateful decoder requires the input queue to be streaming as we
give it buffers, in order to us to get the first resolution change
event, so start streaming the queue during Initialize(). The
stateless decoder doesn't require this behavior, but won't be
affected by it.

BUG=b:149663704
TEST=VDAtest --use_vd passes on Krane.

Change-Id: I783e20c57369bed8e9b8a9a6b9d05a82f14d8a8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235582
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#777337}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62369b5ea314ba964ed2388ee7b5bb25f0507a85