Audio Output not Synchronized to Bluetooth Device
When a bluetooth device is connected that should be the system default or app default we may get a series of device change events that look something like bluetooth device -> system device (previous default) -> bluetooth device.

In these cases if the final transition to bluetooth device occurs within 250ms of the device change notification for the system device we will ignore that change due to the current code that tries to protect against redundant role change notifications.

Since in this case we do want to process the change and transition to the bluetooth device adding an additional check to ensure that the device id has not changed when ignoring a device change due to role timeout.

Bug: 1095667
Change-Id: Ib6518a27f60aa2a0acfa592eb51fffaf375fc407
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245797
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: William Carr <wicarr@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#779943}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b467095b7f6c4af8943ba228313cfe8e66468f5