Reenable SeekWhilePlaying
Currently, MediaRemotingIntegrationTest.SeekWhilePlaying is flaky, and
has been disabled. After looking at the logs, the primary reason is that
media::PipelineIntegrationTestBase::OnEnded is called multiple times,
resulting in a DCHECK and test failure. The stack for the second
OnEnded call is:

media::PipelineImpl::RendererWrapper::CheckPlaybackEnded
media::AudioRendererImpl::Render

Logs:
https://logs.chromium.org/logs/chromium/buildbucket/cr-buildbucket.appspot.com/8890990219267312400/+/steps/media_unittests__with_patch_/0/logs/Flaky_failure:_MediaRemotingIntegrationTest.SeekWhilePlaying__status_CRASH_SUCCESS_/0

Turns out, this is actually a bug in the AudioRendererImpl, where the
old method for determining if the content is ended is flaky, resulting
in multiple calls to OnEnded. This bug has been resolved by a change made on the
AudioRendererImpl, moving to an explicit EndedPlayback signal:

https://chromium-review.googlesource.com/c/chromium/src/+/2021292

Bug: 1043812
Change-Id: I8ca939e5166c3fa3c40f33c7a4e34f3a29d0e06b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258342
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#781482}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: df07149e8b0725ed1bea7045854b6cb10ae241d7