media/gpu/test: Replace memcpy with libyuv::CopyPlane in copying planes
memcpy is used in copying from a plane to the other plane,
strides of which are different from each other, in some places in
media/gpu/test. This CL replaces them with libyuv::CopyPlane
because the  processing is the same as libyuv::CopyPlane.

Bug: None
Test: video.EncodeAccel.* on eve
Test: video_encode_accelerator_tests on eve
Change-Id: Ia57ef4780fd0492dc6ad1a78c5817bee6062ebfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260272
Reviewed-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#781697}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc4bf762e3fa76c247d699b7b4d799beee70ed75