Convert base::CancelableCallback in media::WebMediaPlayerImpl
WebMediaPlayerImpl uses a base::CancelableCallback to hold the callback
that will receive a routing token for the Android overlay. The callback
is canceled if DisableOverlay is called before the token is received.

The type of this callback should be media::RoutingTokenCallback which is
a base::OnceCallback, so we'll switch to base::CancelableOnceCallback.

BUG=1102651

Change-Id: I50845dba9be97f30b86f8ebb71a152457ff3df7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284208
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786446}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: edd780b7d459b5d93a65b4ed419a4efcf351b85c