media/gpu/vaapi_wrapper: make kProfileMap an actual map.
kProfileMap is a map (says the comment) and behaves like one, but it's
implemented as an array of structs, which makes searching cumbersome.
This CL makes it an std::map<>, renames it to kMediaToVAProfileMap,
and updates callsites.

This was prompted because during debugging I saw a pesky, yet persistent,
  "Fall back to constrained baseline profile."
in the logs, that made me look at this area.

No new functionality intended.

Bug: 1040291
Change-Id: I73ce6819bc66d067b39cc2ccb14d88f2830dae18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290690
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787006}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1857f3fe28be1c586bf18455e1a4783f83246bf0