Change CreateCdm to take CdmConfig and return CDM interface
This is in preparation for an upcoming change where we will select a CDM
source in MediaInterfaceProxy based on the CDM config as well as provide
a fallback in the case the CdmFactory is unavailable. The latter case
requires us to have a PendingRemote returned from the call rather than
have a PendingReceiver passed in so we can detect failure in that case.

BUG=b:153111783,chromium:1010341
TEST=Unit tests pass, Widevine content plays back on Chrome OS

Change-Id: I53d16d1bb319a99c191c54e73f29921cf33c09e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283296
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787355}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7075aa9ad99807e50d042d71b719de65b673b25a