[OnionSoup] Move the audio subsystem code to use frame tokens
This is a preparation CL for Onion Souping content/renderer/media/audio.
Audio code uses Frame IDs in many places and across threads in the
renderer process. This makes moving the code to Blink difficult.
This CL switches frame IDs to frame tokens, which are accessible in
Blink and simplify a future migration.

The change to Frame Tokens is largely mechanical, except for the
deregistration of AudioOutputIPCFactories in RenderFrameImpl, which is
moved from the destructor to FrameDetached() since the token is no
longer available in the destructor.

Change-Id: Ib808d7d42c0a252ab4d8210c48d8994db1113ccb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300442
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Antonio Gomes <tonikitoo@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#789512}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6bd37bbf794f957499f9dd9966b65e3360cd14a8