media/gpu/video_decode_accelerator_tests: Use a parser to check if a bitstream has a stream configuration
EncodedDataHelper::HasConfigInfo() checks if a vp8 and vp9 stream
has a stream configuration by checking the first bit of the
stream. This is correct for vp8, but wrong for vp9. Thus it
always returns true in vp9 stream. This bug is not critical
because video_decode_accelerator_tests checks if there is a
stream configuration, but doesn't check the count.

This CL fixes the bug by using parsers for vp8 and vp9 both as
well as for h264.

Bug: None
Test: video_decode_accelerator_tests on atlas
Change-Id: Id1b784e087a102ece40b704d876f0796c0edbc53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2319519
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#792525}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4429914d8c5763101ec345a09c106adf1486c1b