Limit AAudio UTs to supported devices
Currently, when running AAudioOutputStream unit tests, we set a flag,
and leave it up to the AudioManager to choose which Android audio
library it uses. We don't check whether or not we are actually using
AAudio. This can result in running tests use OpenSLES, but are listed as
AAudio tests when they fail.

This CL explicitly checks whether we are using AAudio after setting the
flags, and skips tests that fallback to OpenSLES.

Bug: 1101283
Change-Id: I0e91c11f0f61c13ce1e68061f579b2637fb971ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335485
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#794321}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d7bbdf99ac0399830225c1cacd74f9fa7b4bc4a8