H264Decoder complete DecoderBuffer frames
This adds a flag, which defaults to enabled, which causes
the H264Decoder to assume that a DecoderBuffer contains a complete
picture and will output the frame to the accelerator after
processing the DecoderBuffer, rather than waiting for a signal of
EOS/AUD or the next picture in the bitstream.

The flag is temporary so we can verify if this breaks anything.

BUG=chromium:1105268
TEST=Verified decode latency is reduced with chrome tracing in apprtc

Change-Id: I864337be34634960eacdeaa26d2811d21b8e142d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335338
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#795212}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9340e24644ac91986ca7a58156806ccf9889f577