Add method to WebMediaPlayerClient to get metadata for text tracks
There is an upload warning for use of std::vector, with an attached
note saying that it's acceptable at the blink/chromium layer, which
I think this falls under. This CL should only be merged after the
devtools frontend CL here:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2336204

This also adds a "Chapters" type to the enum used in TextTrackConfig
because apparently that is a thing blink supports.

Bug: 1090022
Change-Id: I19cd27b2c926bef1c36fed19efe8599c9d168386
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337402
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#795465}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1c37a851d6d639da058091744bdf362d9de543e7